requirements.txt
flask==3.0.0
flask-cors==4.0.0
firebase-admin==6.3.0
google-generativeai==0.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
